Skeletal code
Class Vehicle - Keeps track of which lane it is moving in, its current
velocity and position in the lane.
Class Lane - Keeps track of the occupied sites of the lane and the
velocities at occupied sites.
Class Road - Keeps track of multiple lanes, generates traffic and
initiates updates to traffic